Move getAtomName to x11utils
