+++ vstgui/lib/platform/linux/x11utils.cpp
@@ -142,7 +142,25 @@ namespace Atoms {
 Atom xEmbedInfo ("_XEMBED_INFO");
 Atom xEmbed ("_XEMBED");
 
-//------------------------------------------------------------------------
 }
+
+//------------------------------------------------------------------------
+std::string getAtomName (xcb_atom_t atom)
+{
+	std::string name;
+	auto xcb = RunLoop::instance ().getXcbConnection ();
+	auto cookie = xcb_get_atom_name (xcb, atom);
+	if (auto reply = xcb_get_atom_name_reply (xcb, cookie, nullptr))
+	{
+		auto length = xcb_get_atom_name_name_length (reply);
+		name.assign (
+			xcb_get_atom_name_name (reply),
+			xcb_get_atom_name_name_length (reply));
+		free (reply);
+	}
+	return name;
+}
+
+//------------------------------------------------------------------------
 } // X11
 } // VSTGUI
